% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{discover_cli_scripts}
\alias{discover_cli_scripts}
\title{discover available cli scripts}
\usage{
discover_cli_scripts(package = NULL)
}
\arguments{
\item{package}{The name of a specific package to check}
}
\description{
discover available cli scripts
}
