% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_cli_scripts}
\alias{install_cli_scripts}
\title{installed all discovered cli scripts}
\usage{
install_cli_scripts(package = NULL)
}
\arguments{
\item{package}{The name of a specific package to check}
}
\description{
installed all discovered cli scripts
}
