% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_rclikit_script}
\alias{use_rclikit_script}
\title{Create an rclikit compliant script for your package}
\usage{
use_rclikit_script(scriptname, description, short = FALSE)
}
\arguments{
\item{scriptname}{The name of your script}

\item{description}{A one-line description of your script}

\item{short}{use the short template. The default long template provides a broader base upon which to build your own script}
}
\description{
Create an rclikit compliant script for your package
}
